Mise en place de la securité JWT sur project et contributor

+++ monitoring/contributors/views.py
@@ -1,5 +1,7 @@
 from monitoring.contributors.models import Contributor
 from monitoring.contributors.serializers import ContributorSerializer
+from monitoring.permissions import IsContributor
+from monitoring.projects.models import Project
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
@@ -7,7 +9,12 @@ from rest_framework.views import APIView
 
 class ContributorProjectAPIView(APIView):
 
+    permission_classes = [IsContributor]
+
     def post(self, request, project_id):
+        project = Project.objects.get(project=project_id)
+        if project.author_user != request.user:
+            return Response({"message": "Vous n'êtes pas l'auteur du projet."}, status=status.HTTP_401_UNAUTHORIZED)
         serializer = ContributorSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
@@ -18,7 +25,7 @@ class ContributorProjectAPIView(APIView):
 
         try:
             contributors = Contributor.objects.filter(project=project_id)
-            contributor_last_names = [contributor.user_id.last_name for contributor in contributors]
+            contributor_last_names = [contributor.user.last_name for contributor in contributors]
             return Response(contributor_last_names, status=status.HTTP_200_OK)
         except Contributor.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
@@ -28,6 +35,11 @@ class DeleteContributorProjectAPIView(APIView):
 
     def delete(self, request, project_id, user_id):
         try:
+            project = Project.objects.get(project=project_id)
+            print(request.user)
+            print(project.author_user)
+            if project.author_user != request.user:
+                return Response({"message": "Vous n'êtes pas l'auteur du projet."}, status=status.HTTP_401_UNAUTHORIZED)
             contributor = Contributor.objects.get(project_id=project_id, user_id=user_id)
             contributor.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
